server/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
apturl==0.5.2
asgiref==3.7.2
attrs==23.2.0
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.3.2
click==8.0.3
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==3.2.12
django-extensions==3.2.3
django-filter==23.5
django-rest-passwordreset==1.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
exceptiongroup==1.2.0
fastapi==0.110.0
filelock==3.13.1
graphviz==0.20.1
h11==0.14.0
httplib2==0.20.2
httptools==0.6.1
idna==3.3
importlib-metadata==4.6.4
inflection==0.5.1
jeepney==0.7.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
macaroonbakery==1.3.1
Markdown==3.5.2
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
pexpect==4.8.0
Pillow==9.0.1
pipenv==2023.12.1
platformdirs==4.2.0
protobuf==3.12.4
psutil==5.9.0
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
pydantic==2.6.3
pydantic_core==2.16.3
pydot==2.0.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==3.1.2
pyRFC3339==1.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
referencing==0.33.0
reportlab==3.6.8
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.18.0
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
starlette==0.36.3
systemd-python==234
terminator==2.1.1
tweepy==4.14.0
typing_extensions==4.9.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.5
uv==0.1.9
uvicorn==0.27.1
uvloop==0.19.0
virtualenv==20.25.1
virtualenv-clone==0.3.0
wadllib==1.3.6
watchfiles==0.21.0
websockets==12.0
xdg==5
xkit==0.0.0
zipp==1.0.0
